Refactor: WriteWorry screen
