+++ screens/WriteWorry/WorryInputs/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  postTitle: {
+    height: 35,
+    margin: 30,
+    marginBottom: 0,
+    borderRadius: 10
+  },
+  contents: {
+    height: 230
+  }
+});
+
+export default styles;
