+++ screens/WriteWorry/WriteWorry.js
@@ -11,9 +11,9 @@ import { useNavigation } from "@react-navigation/native";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
 
 import Title from "../../components/Title/Title";
-import Picker from "../../components/Picker/Picker";
-import TextInput from "../../components/TextInput/TextInput";
 import AlertModal from "../../components/AlertModal/AlertModal";
+import WorryInputs from "./WorryInputs/WorryInputs";
+import WorryInfoPickers from "./WorryInfoPickers/WorryInfoPickers";
 import WriteWorryButtons from "./WriteWorryButtons/WriteWorryButtons";
 
 import { postNewWorryPost, patchPost } from "../../api/postApi";
@@ -25,13 +25,6 @@ import {
   DETAIL_POST,
   MY_POST_STORAGE
 } from "../../constants/navigationName";
-import {
-  PAIN,
-  LOVE,
-  COURSE,
-  FRIEND,
-  EMPLOYMENT
-} from "../../constants/category";
 import {
   PUBLIC,
   PRIVATE,
@@ -81,42 +74,6 @@ const WriteWorry = ({ route }) => {
 
     return unSubscribe;
   }, [navigation]);
-
-  const postTypes = [
-    { label: PUBLIC, value: PUBLIC },
-    { label: PRIVATE, value: PRIVATE }
-  ];
-  const anonymousTypes = [
-    { label: NICKNAME, value: NICKNAME },
-    { label: ANONYMOUNS, value: ANONYMOUNS }
-  ];
-  const categoryTypes = [
-    { label: PAIN, value: PAIN },
-    { label: LOVE, value: LOVE },
-    { label: COURSE, value: COURSE },
-    { label: FRIEND, value: FRIEND },
-    { label: EMPLOYMENT, value: EMPLOYMENT }
-  ];
-
-  const handlePostPickerChange = (item) => {
-    setPostType(item);
-  };
-
-  const handleAnonymousePickerChange = (item) => {
-    setAnonymousType(item);
-  };
-
-  const handleCategoryPickerChange = (item) => {
-    setCategory(item);
-  };
-
-  const handleWorryContentsChange = (contents) => {
-    setWorryContents(contents);
-  };
-
-  const handlePostTitleChange = (title) => {
-    setPostTitle(title);
-  };
 
   const clearErrorMessage = () => {
     setErrorMessage(null);
@@ -202,36 +159,19 @@ const WriteWorry = ({ route }) => {
                 source={letterPage}
               >
                 <ScrollView>
-                  <TextInput
-                    value={postTitle}
-                    style={styles.postTitle}
-                    placeholder="고민의 제목을 적어주세요"
-                    handleInputChange={handlePostTitleChange}
+                  <WorryInputs
+                    postTitle={postTitle}
+                    worryContents={worryContents}
+                    handleTitleChange={setPostTitle}
+                    handleContentsChange={setWorryContents}
                   />
-                  <TextInput
-                    isMultiline={true}
-                    value={worryContents}
-                    style={styles.contents}
-                    placeholder="고민 거리를 작성해보세요"
-                    handleInputChange={handleWorryContentsChange}
-                  />
-                  <Picker
-                    label="공개 여부"
-                    value={postType}
-                    itemList={postTypes}
-                    handleChange={handlePostPickerChange}
-                  />
-                  <Picker
-                    label="익명 여부"
-                    value={anonymousType}
-                    itemList={anonymousTypes}
-                    handleChange={handleAnonymousePickerChange}
-                  />
-                  <Picker
-                    label="고민 카테고리"
-                    value={category}
-                    itemList={categoryTypes}
-                    handleChange={handleCategoryPickerChange}
+                  <WorryInfoPickers
+                    postType={postType}
+                    category={category}
+                    anonymousType={anonymousType}
+                    handlePostPickerChange={setPostType}
+                    handleCategoryPickerChange={setCategory}
+                    handleAnonymousePickerChange={setAnonymousType}
                   />
                 </ScrollView>
               </ImageBackground>
